fix: jacoco #240603 #13

+++ library-zero-file-manager-kotlin/src/main/kotlin/com/snc/zero/filemanager/file2/FSFile.kt
@@ -38,6 +38,9 @@ class FSFile {
         @JvmStatic
         @Throws(IOException::class)
         fun delete(file: File): Boolean {
+            if (!file.exists()) {
+                return false
+            }
             return file.delete()
         }
